Добавлена заглушка изображений для тренеров Осуществлен сбор направлений из БД на страницу

+++ src/main/java/com/example/ilsan34/clientManager/service/DirectionService.java
@@ -2,16 +2,17 @@ package com.example.ilsan34.clientManager.service;
 
 import com.example.ilsan34.clientManager.model.Direction;
 
+import java.math.BigInteger;
 import java.util.List;
 
 public interface DirectionService {
     void addNewDirection(Direction direction);
 
-    Direction getById(Long id);
+    Direction getById(BigInteger id);
 
     void updateDirection(Direction direction);
 
-    void deleteDirection(Long id);
+    void deleteDirection(BigInteger id);
 
     List<Direction> getAll();
